Added noplatform option to build
git-svn-id: http://subversion.assembla.com/svn/SIM_WIN/trunk@271 e0e46c49-be69-4f5a-ad62-21024a331aea

+++ installer/build.cmd
@@ -2,7 +2,9 @@
 echo Don't forget to install FirefoxPortable and the flash player
 pause
 
+set InnoCC="C:\Program Files (x86)\Inno Setup 5\iscc.exe"
 
+if (%1)==(noplatform) goto noplatform
 rem Clean up first to force complete rebuild - optional
 cd ..\extension\platform
 del /q *.pyc 2> nul:
@@ -13,6 +15,7 @@ c:\python25\python setup.py py2exe
 if errorlevel 1 goto error 
 
 cd ..\..\installer
+:noplatform
 
 rem clean up firefox portable data
 rd /s/q MaavisPortable\App\FirefoxPortable\Data
@@ -20,9 +23,8 @@ mkdir MaavisPortable\App\FirefoxPortable\Data
 
 rem "C:\Program Files\Inno Setup 5\iscc" "Maavis.iss"
 rem if errorlevel 1 goto error 
-"C:\Program Files\Inno Setup 5\iscc" "MaavisPortable.iss"
+%InnoCC% "MaavisPortable.iss"
 if errorlevel 1 goto error 
-
 goto OK
 
 :error
